[Ploobs Gamma Correction] Added More 2 processor (Blur and Load from Texture2D) to this project

+++ PloobsEngine/PloobsEngineContent/Effects/Deferred/InstancedBilboard.fx
@@ -21,10 +21,10 @@ struct BBVertexToPixel
 };
 struct BBPixelToFrame
 {
-    half4 Color : COLOR0;
-    half4 Normal : COLOR1;
-    half4 Depth : COLOR2;
-    half4 Extra1 : COLOR3;
+    float4 Color : COLOR0;
+    float4 Normal : COLOR1;
+    float4 Depth : COLOR2;
+    float4 Extra1 : COLOR3;
 };
 
 //------- Technique: CylBillboard --------
